chore: rename src to packages and delete deps
