+++ packages/browser/index.js
