+++ packages/react-native/.gitattributes
