+++ packages/react-native/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { Platform, StyleSheet, Text, View } from "react-native";
 
-import { getV2BrowserResponse, getV3BrowserResponse } from "../shared/utils";
+import { getV2BrowserResponse, getV3BrowserResponse } from "../utils/utils";
 
 const instructions = Platform.select({
   ios: "Press Cmd+R to reload,\n" + "Cmd+D or shake for dev menu",
