+++ packages/react-native/AppEntry.js
