+++ packages/react-native/__tests__/App.js
