+++ packages/react-native/android/app/proguard-rules.pro
