+++ packages/react-native/android/app/src/main/AndroidManifest.xml
