+++ packages/react-native/android/app/src/main/java/com/expo/MainApplication.java
