+++ packages/react-native/android/app/src/main/res/mipmap-xhdpi/ic_launcher_round.png
