+++ packages/react-native/android/app/src/main/res/values/strings.xml
