+++ packages/react-native/android/app/src/main/res/values/styles.xml
