+++ packages/react-native/android/gradle.properties
