+++ packages/react-native/android/keystores/debug.keystore.properties
