+++ packages/react-native/android/settings.gradle
