+++ packages/react-native/app.json
