+++ packages/react-native/ios/expo.xcworkspace/contents.xcworkspacedata
