+++ packages/react-native/ios/expo/Base.lproj/LaunchScreen.xib
