+++ packages/react-native/ios/expo/Images.xcassets/AppIcon.appiconset/Contents.json
