+++ packages/react-native/ios/expo/Images.xcassets/Contents.json
