+++ packages/react-native/ios/expo/Info.plist
